// src/app/core/_base/layout/models/usermovenco.model.ts
export class UserMovencoModel {
    'userTypeId': string;
    'stationId': number;
    'statusId': string;
    'userName': string;
    'userSurname': string;
    'userLastname': string;
    'userRole': string;
    'userArea': string;
    'userBirthday': string;
    'userMobile': string;
    'expirationDate': string;
    'bloodtype': string;
    'alergiesCondition': string;
    'email': string;
    'street': string;
    'interiorNumber': string;
    'exteriorNumber': string;
    'zipCode': string;
    'neighborhood': string;
    'state': string;
    'city': string;
    'contactName': string;
    'contactSurname': string;
    'contactLastname': string;
    'contactPhone': string;
    'classvehicleId': string;
    'shareBalance': boolean;
    'addUser': boolean;
  }